+++ CHANGELOG.md
@@ -20,6 +20,9 @@ Other Notable Changes:
 * Fix bug in the literal_eval module code used when we need python-2.4 compat
 * Added --ignore-certs, -c option to ansible-galaxy. Allows ansible-galaxy to work behind a proxy
   when the proxy fails to forward server certificates.
+* Fixed bug where tasks marked no_log were showing hidden values in output if
+  ansible's --diff option was used.
+* Fix bug with non-english locales in git and apt modules
 
 ## 1.9.4 "Dancing In the Street" - Oct 10, 2015